+++ config/runtime.exs
@@ -3,7 +3,7 @@ import Config
 config :save_it, :timezone, System.get_env("TZ") || "Asia/Tokyo"
 
 config :save_it, :telegram_bot_token, System.get_env("TELEGRAM_BOT_TOKEN")
-config :ex_gram, token: System.get_env("TELEGRAM_BOT_TOKEN")
+config :ex_gram, token: System.get_env("TELEGRAM_BOT_TOKEN"), adapter: ExGram.Adapter.Req
 
 config :save_it, :cobalt_api_url, System.get_env("COBALT_API_URL", "http://localhost:9001")
 
@@ -30,4 +30,4 @@ config :sentry,
   environment_name: config_env(),
   enable_source_code_context: true,
   root_source_code_path: File.cwd!(),
-  included_dependencies: [:req, :jason, :hackney]
+  included_dependencies: [:req, :jason]